requirements.txt
anthropic
reportlab
PyMuPDF
Pillow
PyPDF2
requests
beautifulsoup4
numpy
json5
python-dotenv
